added gitignore

+++ pria/robot_commander.py
@@ -7,6 +7,7 @@ import math
 import numpy as np
 import os
 import yaml
+from datetime import datetime
 
 import cv2
 from cv_bridge import CvBridge
@@ -45,25 +46,38 @@ class RobotCommander(Node):
         self.image_count_max = self.get_parameter('images_count').get_parameter_value().integer_value
         self.image_count = 0
 
+        self.points = np.empty((self.image_count_max,7),dtype=float)
+
         self.image_height = 240
         self.image_width = 320
 
-        self.parent = os.getcwd()
-        self.path = os.path.join(self.parent,self.directory)
-        self.img_path = os.path.join(self.path, 'imgs')
-        try:
-            os.mkdir(self.path)
-        except FileExistsError:
-            pass
+        self.param_exist = True
+        if self.directory == 'folder':
+            print("Folder parameter not set.")
+            self.param_exist = False
 
-        try:
-            os.mkdir(self.img_path)
-        except FileExistsError:
-            pass
+        if self.image_count_max < 1:
+            print("Image count not set.")
+            self.param_exist = False
 
-        self.gt_path = os.path.join(self.path,'gt.yaml')
-        
-        self.get_logger().info('Dataset folder in {}'.format(self.path))
+        if self.param_exist:
+            self.parent = os.getcwd()
+            self.path = os.path.join(self.parent,self.directory)
+            self.img_path = os.path.join(self.path, 'imgs')
+
+            try:
+                os.mkdir(self.path)
+            except FileExistsError:
+                pass
+
+            try:
+                os.mkdir(self.img_path)
+            except FileExistsError:
+                pass
+
+            self.gt_path = os.path.join(self.path,'gt.yaml')
+            
+            self.get_logger().info('Dataset folder in {}'.format(self.path))
 
         self.tf_subscription = self.create_subscription(
             TFMessage,
@@ -86,10 +100,10 @@ class RobotCommander(Node):
             10)
         self.rgb_subscription  # prevent unused variable warning
         self.bridge = CvBridge()
-        # self.cap = cv2.VideoCapture(0)
-
+        
         self.publisher_ = self.create_publisher(String, '/urscript_interface/script_command', 10)
-        timer_period = 1 # seconds
+
+        timer_period = 0.01 # seconds
         self.timer = self.create_timer(timer_period, self.timer_callback)
         self.i = 0
         self.inMotion = False
@@ -109,6 +123,11 @@ class RobotCommander(Node):
         self.tf_broadcaster = TransformBroadcaster(self)
         self.tf_static_broadcaster = StaticTransformBroadcaster(self)
 
+        # while self.publish_first_pose() < 0:
+        #     print("Waiting for transform.")
+        self.once = True
+
+
     def send_request(self):
         """
         This function uses the SetIO service to put DO1 to Low before starting
@@ -122,32 +141,31 @@ class RobotCommander(Node):
         """
         This is called every ${timer_period} seconds
         """
-        # translation, rotation = self.lookup_transform('base', 'tool0')
-        translation, rotation = self.lookup_transform('base_link_inertia', 'wrist_3_link', True)
+        if self.once:
+            if self.publish_first_pose() == 0:
+                self.once = False
+                self.generate_points()
 
-        if self.image_count_max > self.image_count:
-            if not self.inMotion and self.state == 0:
-                self.state = 1
-                self.handle_next_pose()
+        elif self.param_exist:
+            if self.image_count_max > self.image_count:
+                if not self.inMotion and self.state == 0:
+                    self.handle_next_pose(self.image_count)
+                    self.state = 1
+            else:
+                #go back to initial pose
+                self.back_to_first_pose()
+                print("Finish time ", datetime.now())
+                
+                #save yaml file with ground truth poses
+                with open(self.gt_path, 'w') as file:
+                    yaml.dump(self.gt, file,  default_flow_style=False)
+
+                # and exit
+                raise SystemExit
+            return
+        
         else:
-            #go back to initial pose
-            self.back_to_first_pose()
-            
-            #save yaml file with ground truth poses
-            with open(self.gt_path, 'w') as file:
-                yaml.dump(self.gt, file,  default_flow_style=False)
-
-            # and exit
-            raise SystemExit
-
-        # if self.cv_image.any():
-        #     ret = cv2.imwrite('tes.jpg', self.cv_image)
-        #     self.get_logger().info('Saving image %s' % ret)
-
-        # if not self.inMotion:
-        #     self.handle_next_pose()
-        #     translation, rotation = self.lookup_transform('next_pose', 'base')
-        #     self.send_urscript(translation, rotation)
+            print("doing nothing")
 
     def tf_listener_callback(self, msg):
         """
@@ -182,17 +200,14 @@ class RobotCommander(Node):
         # self.get_logger().info('I heard: "%s"' % msg.height)
         image = self.bridge.imgmsg_to_cv2(msg, 'bgra8')
         # self.color_filter(self.cv_image)
-        if self.state == 2:
-            T, Q = self.lookup_transform('initial_pose', 'wrist_3_link')
-            # filename = '{},{},{},{},{},{},{}.png'.format(T[0],T[1],T[2],R[0],R[1],R[2],R[3])
-            r = Rotations()
-            r.from_quat(Q[0],Q[1],Q[2],Q[3])
-            R = r.as_matrix()
-
+        if self.state == 2 or (self.state == 1 and self.image_count == 0):
+            # Get current transform for the image ground truth
+            T, Q = self.lookup_transform_('initial_pose', 'wrist_3_link')
+            
             self.gt.update({
                 self.image_count: {
                     'translation':T,
-                    'rotation':np.reshape(R, -1).tolist()
+                    'rotation':Q
                 }
             })
             
@@ -201,6 +216,9 @@ class RobotCommander(Node):
             resized = cv2.resize(image, (self.image_width, self.image_height))
             self.capture_and_save_image(resized, os.path.join(self.img_path, filename))
             self.state = 0
+            if self.image_count == 0:
+                print("Starting time ", datetime.now())
+            print("Image ", self.image_count, "/", self.image_count_max, end='\r')
             self.image_count += 1
 
 
@@ -244,7 +262,7 @@ class RobotCommander(Node):
         Returns:
         np.ndarray: The 4x4 transformation matrix
         """
-        x,y,z,w = quaternion
+        x, y, z, w = quaternion
         r = Rotations()
         r.from_quat(x,y,z,w)
         R = r.as_matrix()
@@ -257,81 +275,162 @@ class RobotCommander(Node):
         """
         Publish a static frame with respect to the robot base of the first pose
         """
-        t = TransformStamped()
+        translation, rotation = self.lookup_transform_('base_link_inertia', 'wrist_3_link')
 
-        # Read message content and assign it to
-        # corresponding tf variables
-        t.header.stamp = self.get_clock().now().to_msg()
-        t.header.frame_id = 'base_link_inertia'
-        t.child_frame_id = 'initial_pose'
-        t.transform = self.initial_pose
+        if np.sum(translation) == 0:
+            return -1
 
-        self.tf_static_broadcaster.sendTransform(t)
+        if (self.initial_pose.translation.x == 0 or self.initial_pose.translation.y == 0 or self.initial_pose.translation.z == 0):   
+            self.initial_pose.translation.x = translation[0]
+            self.initial_pose.translation.y = translation[1]
+            self.initial_pose.translation.z = translation[2]
+            r = Rotations()
+            r.from_quat(rotation[0],rotation[1],rotation[2],rotation[3])
+            self.initial_rotvec = r.as_rotvec()
+            self.initial_matrix = self.create_transformation_matrix(rotation, translation)
+            self.gt.update({
+                'initial_pose': {
+                    'translation':translation,
+                    'rotation':rotation,
+                    'total_images': self.image_count_max,
+                    'height':self.image_height,
+                    'width':self.image_width
+                }
+            })
+            
+        self.publish_tf(np.concatenate((translation, rotation)),'base_link_inertia','initial_pose', True)
 
-    def handle_next_pose(self):
-        t = TransformStamped()
+        return 0
 
-        # Read message content and assign it to
-        # corresponding tf variables
-        t.header.stamp = self.get_clock().now().to_msg()
-        t.header.frame_id = 'initial_pose'
-        t.child_frame_id = 'next_pose'
+    def generate_points(self, z_far=10,z_step=1,r_max=100,r_min=100):
 
-        # We create the translation
-        t.transform.translation.x = random.randrange(-90, 90, 1) / 1000
-        t.transform.translation.y = random.randrange(-90, 90, 1) / 1000
-        t.transform.translation.z = 0.0 #random.randrange(0, -300, -1) / 1000 #
+        if self.image_count_max < 1:
+            print("Image count not set")
+            return
 
-        t_next_pose = [t.transform.translation.x,t.transform.translation.y,t.transform.translation.z]
+        zd = z_far / z_step # Interval
+        z_array = np.arange(z_step) * zd + zd / 2
+
+        r_array = r_min + (r_max - r_min) * (np.arange(z_step) / z_step)
+        a = np.power(r_array,2)
+        a_percentage = a / np.sum(a)
+
+        points_index = 0
+
+        images_per_step = self.image_count_max * a_percentage
+        images_per_step = np.round(images_per_step)
+
+        images_per_step[0] += self.image_count_max - np.sum(images_per_step)
+
+
+        for i in range(z_step):
+
+            # print(int(images_per_step[0]))
+
+            for j in range(int(images_per_step[i])):
+                x = random.randrange(-r_array[i], r_array[i], 1) / 1000
+                y = random.randrange(-r_array[i], r_array[i], 1) / 1000
+                z = - random.randrange(z_array[i] - zd / 2, z_array[i] + zd / 2, 1) / 1000
+
+                # Create a quaternion from euler angles
+                r = Rotations()
+                r.from_euler(0, 0, random.randrange(-30, 30, 1) / 100 )
+                q = r.as_quat()
+
+
+                self.points[points_index] = np.array([x,y,z,q[0],q[1],q[2],q[3]])
+                # print(self.points[points_index])
+                self.publish_tf(self.points[points_index], 'initial_pose', 'point_{}'.format(points_index))
+                points_index += 1
+        
+        self.points[-1] = [0.0,0.0,0.0,0.0,0.0,0.0,1.0]
+        self.points = self.points[(-self.points[:,2]).argsort()]
+
+        self.points_list = self.points.tolist()
+        self.points = np.array(self.sort_points(self.points_list))
+
+    def sort_points(self, points):
+        sorted_points = []
+        current_point = self.points_list.pop(0)
+        sorted_points.append(current_point)
+        
+        while self.points_list:
+            closest_point_index = self.find_closest_point_index(current_point[:3], points[:3])
+            current_point = self.points_list.pop(closest_point_index)
+            sorted_points.append(current_point)
+        
+        return sorted_points
+
+    def distance(self,point1, point2):
+        return np.sqrt((point1[0] - point2[0])**2 + (point1[1] - point2[1])**2 + (point1[2] - point2[2])**2)
+
+    def find_closest_point_index(self,current_point, points):
+        min_distance = float('inf')
+        closest_point_index = None
+        for i, point in enumerate(points):
+            if point is not current_point:
+                dist = self.distance(current_point, point)
+                if dist < min_distance:
+                    min_distance = dist
+                    closest_point_index = i
+        return closest_point_index
+
+    def handle_next_pose(self, index):
+        next_pose = self.points[index,:3]
         # self.get_logger().info('offset transform {} {} {}'.format(t_next_pose[0],t_next_pose[1],t_next_pose[2]))
 
-        # Create a quaternion from euler angles
-        r = Rotations()
-        r.from_euler(0, 0, random.randrange(-30, 30, 1) / 100 )
-        q_next_pose = r.as_quat()
-
-        next_pose_matrix = self.create_transformation_matrix(q_next_pose, t_next_pose)
+        next_pose_matrix = self.create_transformation_matrix(self.points[index,3:], next_pose)
         # self.print_transformation_matrix(next_pose_matrix)
 
         base_to_next_matrix = np.dot(self.initial_matrix,next_pose_matrix)
         # self.print_transformation_matrix(base_to_next_matrix)
-        
-        q = q_next_pose
-
-        t.transform.rotation.x = q[0]
-        t.transform.rotation.y = q[1]
-        t.transform.rotation.z = q[2]
-        t.transform.rotation.w = q[3]
-
-        # Send the transformation
-        self.tf_broadcaster.sendTransform(t)
 
         r = Rotations()
         r.from_matrix(base_to_next_matrix[0:3,0:3])
         next_rotation = r.as_rotvec()
 
-        # self.get_logger().info(t_next_pose)
-        # if translation != -1 and rotation != -1:
-        # self.send_urscript(t_next_pose, q_aux)
         self.send_urscript(base_to_next_matrix[:3,3], next_rotation)
-
-    def back_to_first_pose(self):
-        self.send_urscript(self.initial_matrix[:3,3], self.initial_rotvec)
 
     def send_urscript(self, translation, rotation):
         """
         Using the primary interface to send URScripts programs to move the robot
         """
-
         msg = String()
-        # msg.data = """def my_prog():\nset_digital_out(1, True)\nrv=rpy2rotvec([{},{},{}])\nmovej(p[{},{},{},rv[0],rv[1],rv[2]], a=1.2, v=0.25, r=0)\nset_digital_out(1, False)\nend""".format(r,p,y,translation.x,translation.y, translation.z)
-        # msg.data = """def my_prog():\nset_digital_out(1, True)\nrv=rpy2rotvec([{},{},{}])\nmovej(p[{},{},{},rv[0],rv[1],rv[2]], a=1.2, v=0.25, r=0)\nset_digital_out(1, False)\nend""".format(r,p,y,translation[0],translation[1], translation[2])
         msg.data = """def my_prog():\nset_digital_out(1, True)\nmovej(p[{},{},{},{},{},{}], a=1.2, v=0.25, r=0)\nset_digital_out(1, False)\nend""".format(translation[0],translation[1], translation[2], rotation[0], rotation[1], rotation[2])
         self.publisher_.publish(msg)
         # self.get_logger().info(msg.data)
-        # self.i += 1
 
-    def lookup_transform(self, to_frame_rel, from_frame_rel,array=True):
+
+    def back_to_first_pose(self):
+        self.send_urscript(self.initial_matrix[:3,3], self.initial_rotvec)
+
+
+    def publish_tf(self, pose, head, child, static=False):
+        t = TransformStamped()
+
+        # Read message content and assign it to
+        # corresponding tf variables
+        t.header.stamp = self.get_clock().now().to_msg()
+        t.header.frame_id = head
+        t.child_frame_id = child       
+        
+        # We create the translation
+        t.transform.translation.x = pose[0]
+        t.transform.translation.y = pose[1]
+        t.transform.translation.z = pose[2]
+
+        t.transform.rotation.x = pose[3]
+        t.transform.rotation.y = pose[4]
+        t.transform.rotation.z = pose[5]
+        t.transform.rotation.w = pose[6]
+
+        # Send the transformation
+        if static:
+            self.tf_static_broadcaster.sendTransform(t)
+        else:
+            self.tf_broadcaster.sendTransform(t)
+
+    def lookup_transform_(self, to_frame_rel, from_frame_rel):
         """
         Search for transformations between two specific frames.
         
@@ -348,50 +447,13 @@ class RobotCommander(Node):
             translation = [t.transform.translation.x, t.transform.translation.y,t.transform.translation.z]
             rotation = [t.transform.rotation.x, t.transform.rotation.y, t.transform.rotation.z, t.transform.rotation.w]
 
-            #todo: put this somewhere else
-            if (self.initial_pose.translation.x == 0 or self.initial_pose.translation.y == 0 or self.initial_pose.translation.z == 0):   
-                self.initial_pose = t.transform
-                r = Rotations()
-                r.from_object(t.transform.rotation)
-                self.initial_rotvec = r.as_rotvec()
-                self.initial_matrix = self.create_transformation_matrix(rotation, translation)
-                self.publish_first_pose()
-                self.gt.update({
-                    'initial_pose': {
-                        'translation':translation,
-                        'rotation':rotation,
-                        'total_images': self.image_count_max,
-                        'height':self.image_height,
-                        'width':self.image_width
-                    }
-                })
-                # with open('./test.yaml', 'w') as file:
-                #     yaml.dump(self.gt, file,  default_flow_style=False)
-            #
-
-            if array:
-                return translation, rotation
-            else:
-                return t.transform.translation, t.transform.rotation
+            return translation, rotation
 
         except TransformException as ex:
             self.get_logger().info(
                 f'Could not transform {to_frame_rel} to {from_frame_rel}: {ex}')
-            return -1, -1
-    
-    def print_transformation_matrix(self, H):
-        a, b, c, d = H[0,:]
-        e, f, g, h = H[1,:]
-        i, j, k, l = H[2,:]
-        m, n, o, p = H[3,:]
-        self.get_logger().info('transformation matrix: \n [[{},{},{},{}],\n[{},{},{},{}],\n[{},{},{},{}],\n[{},{},{},{}]]'.format(a,b,c,d,e,f,g,h,i,j,k,l,m,n,o,p))
-
-    
-    def print_rotation_matrix(self, R):
-        a, b, c = R[0,:]
-        d, e, f = R[1,:]
-        g, h, i = R[2,:]
-        self.get_logger().info('rotation matrix: \n [[{},{},{}],\n[{},{},{}],\n[{},{},{}]]'.format(a,b,c,d,e,f,g,h,i))
+            return [0.0,0.0,0.0], [0.0,0.0,0.0,1.0]
+        
 
 def main(args=None):
     rclpy.init(args=args)
